+++ playground/with-cdn/vite.config.js
@@ -1,65 +1,63 @@
-import { defineConfig } from 'vite'
-import { dirname, resolve } from 'path';
+import { defineConfig } from "vite";
+import { dirname, resolve } from "path";
 
-import symfonyPlugin from 'vite-plugin-symfony';
+import symfonyPlugin from "vite-plugin-symfony";
 import vuePlugin from "@vitejs/plugin-vue";
-import reactPlugin from '@vitejs/plugin-react';
-import { fileURLToPath } from 'url';
+import reactPlugin from "@vitejs/plugin-react";
+import { fileURLToPath } from "url";
 
 const basicPlaygroundDir = dirname(fileURLToPath(import.meta.url));
-const sharedDir = resolve(basicPlaygroundDir, '../../shared')
+const sharedDir = resolve(basicPlaygroundDir, "../../shared");
 
-export default defineConfig(({mode }) => {
+export default defineConfig(({ mode }) => {
   return {
     plugins: [
       vuePlugin(),
       reactPlugin(),
       symfonyPlugin({
-        sriAlgorithm: "sha256"
+        sriAlgorithm: "sha256",
       }),
     ],
-  
-    base: mode === 'development' ? '/build/' : 'http://cdn.with-cdn.symfony-vite-dev.localhost',
-  
+
+    base:
+      mode === "development"
+        ? "/build/"
+        : "http://cdn.with-cdn.symfony-vite-dev.localhost",
+
     publicDir: false,
-  
+
     build: {
+      assetsInlineLimit: 0,
+      outDir: "./public/build",
       manifest: true,
       rollupOptions: {
         input: {
-          "pageAssets": "./assets/page/assets/index.js",
-          "pageImports": "./assets/page/imports/index.js",
-          "pageVue": "./assets/page/vue/main.js",
-          "pageReact": "./assets/page/react/main.jsx",
-          "pageWelcome": "./assets/page/welcome/index.js",
-  
-          "app": "./assets/app.js",
-          "theme": "./assets/theme.scss"
-        }
+          pageAssets: "./assets/page/assets/index.js",
+          pageImports: "./assets/page/imports/index.js",
+          pageVue: "./assets/page/vue/main.js",
+          pageReact: "./assets/page/react/main.jsx",
+          pageWelcome: "./assets/page/welcome/index.js",
+
+          app: "./assets/app.js",
+          theme: "./assets/theme.scss",
+        },
       },
-  
+
       minify: false,
     },
-  
+
     server: {
       // origin: 'http://localhost:5173',
       fs: {
-        allow: [
-          '.',
-          sharedDir
-        ]
-      }
+        allow: [".", sharedDir],
+      },
     },
-  
+
     resolve: {
       alias: {
-        '~': resolve(basicPlaygroundDir, 'assets'),
-        '~shared': sharedDir
-      }
+        "~": resolve(basicPlaygroundDir, "assets"),
+        "~shared": sharedDir,
+      },
     },
-  
-  
-  }
+  };
 });
-
-