+++ src/vite-plugin-symfony/src/index.ts
@@ -1,7 +1,7 @@
 import { Plugin } from "vite";
 import symfonyEntrypoints from "./entrypoints";
-import symfonyStimulus from "./stimulus";
 import symfonyFosRouting from "./fos-routing";
+import symfonyStimulus from "./stimulus/node";
 
 import { VitePluginSymfonyOptions } from "./types";
 import { resolvePluginOptions } from "./pluginOptions";