+++ src/vite-plugin-symfony/src/pluginOptions.ts
@@ -1,6 +1,5 @@
 import { join } from "node:path";
 import {
-  VitePluginSymfonyEntrypointsOptions,
   VitePluginSymfonyFosRoutingOptions,
   VitePluginSymfonyOptions,
   VitePluginSymfonyPartialOptions,
@@ -8,23 +7,9 @@ import {
 } from "./types";
 import { deepMerge } from "~/fos-routing/utils";
 
+import { trimSlashes } from "./entrypoints/utils";
+
 export function resolvePluginOptions(userConfig: VitePluginSymfonyPartialOptions = {}): VitePluginSymfonyOptions {
-  if (typeof userConfig.publicDirectory === "string") {
-    userConfig.publicDirectory = userConfig.publicDirectory.trim().replace(/^\/+/, "").replace(/\/+$/, "");
-
-    if (userConfig.publicDirectory === "") {
-      throw new Error("vite-plugin-symfony: publicDirectory must be a subdirectory. E.g. 'public'.");
-    }
-  }
-
-  if (typeof userConfig.buildDirectory === "string") {
-    userConfig.buildDirectory = userConfig.buildDirectory.trim().replace(/^\/+/, "").replace(/\/+$/, "");
-
-    if (userConfig.buildDirectory === "") {
-      throw new Error("vite-plugin-symfony: buildDirectory must be a subdirectory. E.g. 'build'.");
-    }
-  }
-
   if (typeof userConfig.servePublic === "undefined") {
     userConfig.servePublic = "public";
   }
@@ -85,13 +70,11 @@ export function resolvePluginOptions(userConfig: VitePluginSymfonyPartialOptions
   }
 
   return {
-    buildDirectory: userConfig.buildDirectory,
     debug: userConfig.debug === true,
     enforcePluginOrderingPosition: userConfig.enforcePluginOrderingPosition === false ? false : true,
     enforceServerOriginAfterListening: userConfig.enforceServerOriginAfterListening === false ? false : true,
     exposedEnvVars: userConfig.exposedEnvVars ?? ["APP_ENV"],
     originOverride: userConfig.originOverride ?? null,
-    publicDirectory: userConfig.publicDirectory,
     refresh: userConfig.refresh ?? false,
     servePublic: userConfig.servePublic,
     sriAlgorithm: userConfig.sriAlgorithm ?? false,
@@ -101,31 +84,13 @@ export function resolvePluginOptions(userConfig: VitePluginSymfonyPartialOptions
   };
 }
 
-export function resolveBase(config: VitePluginSymfonyEntrypointsOptions): string {
-  if (typeof config.buildDirectory !== "undefined") {
-    return "/" + config.buildDirectory + "/";
-  }
-  return "/build/";
-}
+export function resolveOutDir(unknownBase: string): string {
+  const baseURL = new URL(unknownBase, import.meta.url);
 
-export function resolveOutDir(config: VitePluginSymfonyEntrypointsOptions): string {
-  let publicDirectory = "public";
-  let buildDirectory = "build";
-  if (typeof config.publicDirectory !== "undefined") {
-    publicDirectory = config.publicDirectory;
-  }
-  if (typeof config.buildDirectory !== "undefined") {
-    buildDirectory = config.buildDirectory;
-  }
-  return join(publicDirectory, buildDirectory);
-}
+  const base = baseURL.protocol === "file:" ? unknownBase : baseURL.pathname;
+  const publicDirectory = "public";
 
-export function resolvePublicDir(config: VitePluginSymfonyEntrypointsOptions) {
-  if (typeof config.publicDirectory !== "undefined") {
-    return config.publicDirectory;
-  }
-
-  return config.servePublic === false ? null : config.servePublic;
+  return join(publicDirectory, trimSlashes(base));
 }
 
 export const refreshPaths = ["templates/**/*.twig"];